Central billing profile fixture
