+++ src/billing/useWorkspaceBillingProfile.test.ts
@@ -1,5 +1,6 @@
-import { Billing, BillingContract } from 'src/libs/ajax/Billing';
+import { Billing, BillingContract, BillingProfile } from 'src/libs/ajax/Billing';
 import { WorkspaceManagerResources, WorkspaceManagerResourcesContract } from 'src/libs/ajax/WorkspaceManagerResources';
+import { azureBillingProfile } from 'src/testing/billing-profile-fixtures';
 import { asMockedFn, mockNotifications, renderHookInActWithAppContexts } from 'src/testing/test-utils';
 
 import { useWorkspaceBillingProfile } from './useWorkspaceBillingProfile';
@@ -10,29 +11,10 @@ jest.mock('src/libs/ajax/WorkspaceManagerResources', () => ({ WorkspaceManagerRe
 describe('useWorkspaceBillingProfile', () => {
   it('returns billing profile', async () => {
     // Arrange
-    const billingProfile = {
-      id: '04cb122b-89a7-4b92-ad69-35f9877406ef',
-      biller: 'direct',
-      displayName: 'TestBillingProfile',
-      description: 'This is only a test.',
-      cloudPlatform: 'AZURE',
-      tenantId: 'dcdfd729-0a05-45e0-bfc9-a2458f8ee29b',
-      subscriptionId: 'aff128c1-8e0c-442a-b5d7-ff435e621fc9',
-      managedResourceGroupId: 'TestMRG',
-      createdDate: '2024-05-28T18:29:26.333295Z',
-      lastModified: '2024-05-28T18:29:26.333295Z',
-      createdBy: 'user@example.com',
-      policies: {
-        inputs: [
-          {
-            namespace: 'terra',
-            name: 'protected-data',
-            additionalData: [],
-          },
-        ],
-      },
+    const billingProfile: BillingProfile = {
+      ...azureBillingProfile,
       organization: {
-        enterprise: false,
+        ...azureBillingProfile.organization,
         limits: {
           autopause: '30',
           persistentdisk: '32',
